Abstract board methods and rspec tests into Board class
