+++ lib/board.rb
@@ -1,3 +1,22 @@
+# require 'game'
 class Board
-  
+  def initialize
+    @board = [' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ']
+  end
+
+  def display_board
+    puts " #{@board[0]} | #{@board[1]} | #{@board[2]} "
+    puts '-----------'
+    puts " #{@board[3]} | #{@board[4]} | #{@board[5]} "
+    puts '-----------'
+    puts " #{@board[6]} | #{@board[7]} | #{@board[8]} "
+  end
+
+  def position_taken?(index)
+    @board[index] != ' '
+  end
+
+  def mark_board(index, char = 'X')
+    @board[index] = char
+  end
 end
